+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## [0.0.2] - 2025-06-26
 
+### Added
+- Rule 0 challenge: Scoutmaster targets the closest person if someone dies
+
 ### Changed
 
 - Switch to Peak mod template